Move baseKeybase to a separate file
